fix 'compress' semantics
The type and semantics of the 'compress' keyword argument were ambiguous. In
the joblib/memory.py module the type was listed as boolean in the docstring and as
default value for the keyword arguments. However, in joblib/numpy_pickle.py it
turned into an integer which eventually becomes a compression level for zlib.
This never showed up anywhere I guess, since by virtue of duck typing False was
interpreted correctly to mean a compression level of 0 and True to mean 1.

Make the 'compress' setting uniform across joblib/memory.py and
joblib/numpy_pickle.py. Adapt the docstrings and the keyword defaults to be
integers all the way.