FIX #146: Heisen test failure caused by thread-unsafe Python lists
This fix uses a queue.Queue datastructure in the failing test. This
datastructure is thread-safe thanks to an internal Lock. This Lock instance is
not picklable hence cause the picklability check of delayed to check fail.

When using the threading backend, picklability is no longer required, hence
this PRs give the user the ability to disable it on a case by case basis.