ENH: Process filter arguments when getting key
This commit adds functionality to joblib.func_inspect.filter_args to have
not only a list of arguments that are ignored, but also a dictionary of
arguments processed in a particular way.  This can be useful when we want
to avoid values that cause a failure to calculate a key.  One example
would be an ndarray containing datetime64 (see issue #188).  This commit
adds the appropriate functionality to the MemorizedFunc and Memory
classes, by adding an attribute `process` analogous to the existing
attribute `ignore`.

Also add a test to see if this works, to
joblib.test.test_func_inspect.test_filter_args.