Fix joblib.test.test_hashing.test_dtype random failure
with python 3.

Rather than trying to pickle numpy.dtype objects, which
give rise to multiple problems with hashing, we used its .desr
attribute which is a full description of the dtype.

This allows not to rely on the persistent_id workaround. The reason it
wasn't working in the first place is that because joblib.dumps(dt) is
not stable across python runs, because the order of dt.fields
changes (dict order randomization in python3).