FIX joblib dump for numpy arrays of dtype object
For object dtype, we do not try to save the numpy array in an auxiliary
file but just do a pickle.save in the master pickle file.

Added a numpy array with dtype object in
test_compressed_pickle_dump_and_load. Also test the code branch where
the array size is bigger than the 'cache_size' argument of joblib.dump.