Add `allow_override` kwarg to Parallel
If true, wrapping a call to `Parallel` with a `parallel_backend`
contextmanager will use the active backend instead of the backend
specified in the `Parallel` call. Default is False (existing behavior).

This allows overriding what backend is used inside libraries that depend
on joblib (e.g. scikit-learn), and is motivated by the
`dask.distributed` joblib backend.

This also changes the kwarg default for `n_jobs` to None to allow for
checking if the user specified a number in the call to `Parallel` (which
is then used), otherwise the global default is used. This is better than
the existing logic, which checked if it was `1`, which meant that if no
backend was specified but `n_jobs` was set to `1`,
`n_jobs=DEFAULT_N_JOBS` would be used instead. The default behavior
remains the same as before.