Switch default n_jobs handling
Now ``allow_override`` dictates if both the backend and n_jobs can be
overridden. This required changing the default value of ``n_jobs`` for
``parallel_backend``, but that should be fine as the docs provide a
warning that it may change at any time.

This also changes the defaults for ``backend`` and ``n_jobs`` to ignore
``parallel_backend`` if not provided. This is more consistent, as the
``parallel_backend`` contextmanager is now always ignored unless
``allow_override=True``.