FIX python-lz4 version comparison
The python-lz4 package has its version string prefixed with "v".
As a result on nixos a test fails.

This commit changes to pkg_resources.parse_version which normalizes
the two schemas correctly and seems to be the recommended way to
compare versions in python at the moment.

Nixos build log https://hydra.nixos.org/build/88820470/nixlog/2.