created a joblib logger.
using logging.(debug,warning, etc) from within a library is unwise as it forces the end user to have a stream handler on the root logger. Now you'll be logging to the joblib namespace, so that users can specify a logging level via `logging.getLogger('joblib').setLevel(logging.DEBUG)` or whatever they want.