Use `inspect.getsourcelines` to get code
This function was not used previously due to a bug in how it cached source
code. This bug was fixed in Python 3.5 which is now EOL. This also allows
us to detect the actual line in source code of a lambda reducing possible
collisions.

See https://bugs.python.org/issue1218234