FIX _get_items_to_delete raising error when items list empty
when calling `.reduce_size` on a newly created memory object without any cached items, it raises a `ValueError: min() arg is an empty sequence`. The argument of `age_limit` must not be None for this case.

Fixed by providing a default value to the `min` function used in calculating the oldest item, which would be the current datetime as no items exist.